+++ src/main/java/gogitek/restaurentorder/service/CategoryService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface CategoryService {
     List<Category> getListCategory();
-    Optional<Category> getCategoryById(int id);
+    Optional<Category> getCategoryById(Long id);
     boolean addCategory(Category category);
-    boolean deleteCategory(int id);
-    void updateCategory(int id, Category category);
+    boolean deleteCategory(Long id);
+    void updateCategory(Long id, Category category);
 }
